+++ src/App.css
@@ -7,16 +7,54 @@
 }
 
 body {
-  background-color: burlywood;
-  display: grid;
-  place-content: center;
   font-size: 1.3rem;
   height: 100vh;
 }
 
 .container {
-  height: 65vh;
-  width: 140vh;
+  width: 100%;
+  height: 90vh;
+  margin-top: 85px;
+}
+
+.navbar {
+  width: 100%;
+  height: 85px;
+  background-color: rgb(99, 99, 99);
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-left: 120px;
+  padding-right: 120px;
+}
+
+.mainpage {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.nav-links {
+  display: flex;
+  justify-content: space-between;
+  width: 25%;
+}
+
+.nav-links * {
+  color: black;
+  font-weight: bold;
+  text-decoration: none;
+}
+
+/*
+.container {
+  height: 100%;
+  width: 100%;
   display: flex;
   justify-content: space-between;
   background-color: rgb(134, 132, 132);
@@ -25,85 +63,4 @@ body {
   -moz-box-shadow: 7px 7px 5px 0 rgba(50, 50, 50, 0.75);
   box-shadow: 7px 7px 5px 0 rgba(50, 50, 50, 0.75);
 }
-
-.quotes-container {
-  height: 100%;
-  width: 49%;
-  margin-right: 2%;
-  margin-left: 2%;
-  display: flex;
-}
-
-.quotes-wrapper {
-  padding: 50px 10px;
-  height: 90%;
-  width: 90%;
-  background: rgb(220, 220, 220);
-  align-self: center;
-}
-
-.quotes-wrapper > * {
-  margin-top: 15px;
-}
-
-.wrapper {
-  display: grid;
-  height: 100%;
-  width: 49%;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(6, 1fr);
-  align-self: self-end;
-}
-
-.screen {
-  grid-column: 1/-1;
-  color: white;
-  text-align: end;
-  padding-right: 10px;
-  padding-top: 18px;
-  font-size: 2rem;
-  overflow: hidden;
-  background-color: rgba(134, 132, 132, 0.703);
-  border: 2px solid rgb(101, 99, 99);
-  border-top-right-radius: 8px;
-}
-
-.btn {
-  display: grid;
-  place-content: center;
-  background-color: rgb(220, 220, 220);
-  color: black;
-  cursor: pointer;
-  border-radius: 4px;
-  font-weight: bold;
-  border: 1px solid rgb(185, 185, 185);
-}
-
-.btn:hover {
-  background-color: gray;
-}
-
-.light-gray {
-  background-color: rgb(220, 220, 220);
-  color: black;
-}
-
-.light-gray:hover {
-  background-color: white;
-}
-
-.orange {
-  background-color: rgba(255, 166, 0, 0.842);
-}
-
-.orange:hover {
-  background-color: white;
-  color: orange;
-}
-
-.zero {
-  grid-column: 1/span 2;
-  place-content: unset;
-  align-items: center;
-  padding-left: 2rem;
-}
+*/
