+++ src/components/Calculator/logic/calculate.js
