+++ src/components/Calculator/logic/operate.js
