+++ src/components/Home/HomePage.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import './HomePage.css';
+
+const Hompage = () => (
+  <div className="homepage-container">
+    <h1>Welcome to our page</h1>
+    <p>
+      Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempore est magnam labore
+      praesentium saepe quaerat ipsam laudantium tempora quisquam cum nesciunt in natus
+      aliquid magni, blanditiis, sed repellat! Voluptatum, similique?
+      Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempore est magnam labore
+      praesentium saepe quaerat ipsam laudantium tempora quisquam cum nesciunt in natus
+      aliquid magni, blanditiis, sed repellat! Voluptatum, similique?
+    </p>
+    <br />
+    <p>
+      Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempore est magnam labore
+      praesentium saepe quaerat ipsam laudantium tempora quisquam cum nesciunt in natus
+      aliquid magni, blanditiis, sed repellat! Voluptatum, similique?
+      Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempore est magnam labore
+      praesentium saepe quaerat ipsam laudantium tempora quisquam cum nesciunt in natus
+      aliquid magni, blanditiis, sed repellat! Voluptatum, similique?
+      Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempore est magnam labore
+      praesentium saepe quaerat ipsam laudantium tempora quisquam cum nesciunt in natus
+      aliquid magni, blanditiis, sed repellat! Voluptatum, similique?
+    </p>
+  </div>
+);
+
+export default Hompage;
